ChatGPTWrapped/requirements.txt
streamlit>=1.52.1
pandas>=2.1
numpy>=1.26
plotly>=5.20
python-dateutil>=2.9
tzdata>=2024.1
